Added Partial Maze representation
Now I can represent any square maze with just walls. The Maze class will need a good deal of features
in the future. I will need a method to directly get and set walls based off of player position, which will
pretty much be a mapping of a position to four Boolean values that represent the walls. I want the
get and set to be very easy to do. A maze generation algorithm should flow very nicely from that.