Algorithm Animation
This is the dirtiest, hackiest piece of code that I have written for this project yet. It has a ton of repetition and there is not much modularity
to it.

Anyway, I created an animation functionality for the depth-first generator. It looks so cool. I want to try out some more maze algorithms
because of this. Very rewarding.

Next steps: Create drop-down menu for algorithms. Include an animate option for a given algorithm. Create hot-keys for easy maze
making.