Finishing Cleaning
Pretty much all cleaning of the code has been done. I have been putting off implementing the algorithm selection menu. I need to do
that next. I still have to follow up on the ease-of-use hot keys as well.

I have some more maze algorithms to put into place. There are two that I have on the top of my head, one of which is very easy
to write. Both will be a cinch to animate.