Added crude hotkeys
<enter> now generates a maze on the spot. ctrl-w and ctrl-q now have behaviour for exiting windows and quitting out of the program.
Given the manner in which mazes are animated the ctrl-w/q will not register with tkinter until an animation is over. It's not ideal but it
is minor and I will overlook this for now. In the future I may want to fix this and enable an animation to be paused/resumed with the space
bar.

I will now create the drop down menu for algorithms and solidify the algorithm selection process within the code.

Following that I will put in one (last?) maze algorithm that also comes with an animation. It will be a binary tree maze. The animation
for it will hopefully look as cool as the depth-first animation.