Optimization Attempt
I took a stab at switching the priority queue implementation from a dynamically sized list to a binary heap, so as to improve Prim's Algo's
performance. This switch did not work well. Prim's speed stayed the same, if not slower. This is a little bit of a failure in that I did not
analyze the properties of the operations that I was asking this priority queue to complete. In short, since I request the queue to remove
all edges that are no longer possible after each move, I am adding an extra amount of complexity to the remove process. 

I will evaluate the best way to go about this and actually look at what I need to optimize this time. I am thinking the dynamically sized
list will be able to accommodate me and that I just need to help it out with some of its operations.