Splitting generator.py in two. Refactoring Algorithms. Doesn't function.
FYI, this commit will not function because the Algorithms int MazeGenerators.py
currently make calls to methods in app.py. However, there will be a ton
of edits coming down the line, so here's to breaking up the commits.

My intention is for app.py to (for the most part) contain all things that
relate to the graphics, and the U-I. The goal is for it to have minimal
interweaving with the maze generators. Secondly, I want to eliminate all
animation functionality from the maze generators. This will make the code
more modular and clean. Maze Generation algorithms should not have to
worry about animation in the slightest.

As shown in this commit, I am halfway through transitioning the maze
generators into something I call 'step-form'. Every single one of my
algorithms works by adding one node to the tree that is the maze
one at a time. So, each algorithm will have a step method that adds
one square to the maze and also returns the edge that was traversed
by the method. With just information on the edge, an App object from
app.py will be able to draw the cell being added to the maze. Easy.