Generation Algorithm Re-Vamp (does not run)
This commit will not run because it is dependent on the commit that
will immediately follow it, i.e.: the corresponding changes to app.py
that will allow the mazes to actually generate and animate with the
new underlying MazeGenerationAlgorithm objects.

On to the details.

All of the maze generation algorithms now fall under the MGAlgorithm
class. A child of MGAlgorithm must have a step() method. Thats it.
Because of this change, I have restructured the inner workings of
each of the algorithms. Essentially, each algorithm is an object
and keeps track of its internal state. with each call of step() it
will mutate its maze and return a Node object representing where it
decided to create a connecting corridor in the maze.

Repeated calling of the step() method is definitely less efficient
than just iterating through the algorithm and having inline animation
commands, but this makes the code much more modular and clean. Also,
the small loss in speed is not a big deal, because this program is NOT
meant to be optimal. Rather, it is meant to be reasonable and FUN.

I also made some small changes to the Node class and some other helper
functions. The changes are meant to improve readability and consistency
across the program.

Continue on to the next commit to get the remainder of this initial
refactor.